backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
playwright==1.39.0
openai==1.3.0
nest-asyncio==1.5.8
pandas==2.1.1
pillow==10.0.1
